backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
httpx==0.27.2
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
openai==1.51.0
requests==2.31.0
python-multipart==0.0.9
pypdf2==3.0.1